--- a/extension/parquet/parquet_file_metadata_cache.cpp
+++ b/extension/parquet/parquet_file_metadata_cache.cpp
@@ -20,8 +20,36 @@ string ParquetFileMetadataCache::GetObjectType() {
 }
 
 bool ParquetFileMetadataCache::IsValid(CachingFileHandle &new_handle) const {
-	return ExternalFileCache::IsValid(validate, version_tag, last_modified, new_handle.GetVersionTag(),
-	                                  new_handle.GetLastModifiedTime());
+	return ExternalFileCache::IsValid(validate, new_handle.GetVersionTag(), new_handle.GetLastModifiedTime(),
+	                                  version_tag, last_modified);
+}
+
+ParquetCacheValidity ParquetFileMetadataCache::IsValid(const OpenFileInfo &info) const {
+	if (!info.extended_info) {
+		return ParquetCacheValidity::UNKNOWN;
+	}
+	auto &open_options = info.extended_info->options;
+	const auto validate_entry = open_options.find("validate_external_file_cache");
+	if (validate_entry != open_options.end()) {
+		// check if always valid - if so just return valid
+		if (BooleanValue::Get(validate_entry->second)) {
+			return ParquetCacheValidity::VALID;
+		}
+	}
+	const auto lm_entry = open_options.find("last_modified");
+	if (lm_entry == open_options.end()) {
+		return ParquetCacheValidity::UNKNOWN;
+	}
+	time_t new_last_modified = Timestamp::ToTimeT(lm_entry->second.GetValue<timestamp_t>());
+	string new_etag;
+	const auto etag_entry = open_options.find("etag");
+	if (etag_entry != open_options.end()) {
+		new_etag = StringValue::Get(etag_entry->second);
+	}
+	if (ExternalFileCache::IsValid(false, new_etag, new_last_modified, version_tag, last_modified)) {
+		return ParquetCacheValidity::VALID;
+	}
+	return ParquetCacheValidity::INVALID;
 }
 
 } // namespace duckdb