--- a/src/common/file_system.cpp
+++ b/src/common/file_system.cpp
@@ -368,8 +368,12 @@ string FileSystem::ExpandPath(const string &path) {
 // LCOV_EXCL_START
 unique_ptr<FileHandle> FileSystem::OpenFileExtended(const OpenFileInfo &path, FileOpenFlags flags,
                                                     optional_ptr<FileOpener> opener) {
-	// for backwards compatibility purposes - default to OpenFile
 	throw NotImplementedException("%s: OpenFileExtended is not implemented!", GetName());
+}
+
+bool FileSystem::ListFilesExtended(const string &directory, const std::function<void(OpenFileInfo &info)> &callback,
+                                   optional_ptr<FileOpener> opener) {
+	throw NotImplementedException("%s: ListFilesExtended is not implemented!", GetName());
 }
 
 unique_ptr<FileHandle> FileSystem::OpenFile(const string &path, FileOpenFlags flags, optional_ptr<FileOpener> opener) {
@@ -389,6 +393,10 @@ unique_ptr<FileHandle> FileSystem::OpenFile(const OpenFileInfo &file, FileOpenFl
 }
 
 bool FileSystem::SupportsOpenFileExtended() const {
+	return false;
+}
+
+bool FileSystem::SupportsListFilesExtended() const {
 	return false;
 }
 
@@ -483,9 +491,46 @@ void FileSystem::RemoveDirectory(const string &directory, optional_ptr<FileOpene
 	throw NotImplementedException("%s: RemoveDirectory is not implemented!", GetName());
 }
 
+bool FileSystem::IsDirectory(const OpenFileInfo &info) {
+	if (!info.extended_info) {
+		return false;
+	}
+	auto entry = info.extended_info->options.find("type");
+	if (entry == info.extended_info->options.end()) {
+		return false;
+	}
+	return StringValue::Get(entry->second) == "directory";
+}
+
 bool FileSystem::ListFiles(const string &directory, const std::function<void(const string &, bool)> &callback,
                            FileOpener *opener) {
+	if (SupportsListFilesExtended()) {
+		return ListFilesExtended(
+		    directory,
+		    [&](const OpenFileInfo &info) {
+			    bool is_dir = IsDirectory(info);
+			    ;
+			    callback(info.path, is_dir);
+		    },
+		    opener);
+	}
 	throw NotImplementedException("%s: ListFiles is not implemented!", GetName());
+}
+
+bool FileSystem::ListFiles(const string &directory, const std::function<void(OpenFileInfo &info)> &callback,
+                           optional_ptr<FileOpener> opener) {
+	if (SupportsListFilesExtended()) {
+		return ListFilesExtended(directory, callback, opener);
+	} else {
+		return ListFiles(directory, [&](const string &path, bool is_dir) {
+			OpenFileInfo info(path);
+			if (is_dir) {
+				info.extended_info = make_shared_ptr<ExtendedOpenFileInfo>();
+				info.extended_info->options["type"] = "directory";
+			}
+			callback(info);
+		});
+	}
 }
 
 void FileSystem::MoveFile(const string &source, const string &target, optional_ptr<FileOpener> opener) {