--- a/src/common/local_file_system.cpp
+++ b/src/common/local_file_system.cpp
@@ -1181,7 +1181,7 @@ bool LocalFileSystem::ListFilesExtended(const string &directory,
 		// file size
 		int64_t file_size_bytes =
 		    (static_cast<int64_t>(ffd.nFileSizeHigh) << 32) | static_cast<int64_t>(ffd.nFileSizeLow);
-		options.emplace("file_size", Value::BIGINT(file_size_bytes);
+		options.emplace("file_size", Value::BIGINT(file_size_bytes));
 		// last modified time
 		auto last_modified_time = FiletimeToTimeT(ffd.ftLastWriteTime);
 		options.emplace("last_modified", Value::TIMESTAMP(Timestamp::FromTimeT(last_modified_time)));