--- a/src/common/types/timestamp.cpp
+++ b/src/common/types/timestamp.cpp
@@ -562,19 +562,22 @@ time_t Timestamp::ToTimeT(timestamp_t timestamp) {
 }
 
 timestamp_t Timestamp::FromTimeT(time_t time) {
-	struct tm tm {};
 #ifdef DUCKDB_WINDOWS
-	tm = localtime(&tm);
+	auto tm = localtime(&time);
 #else
-	localtime_r(&time, &tm);
+	struct tm tm_storage {};
+	auto tm = localtime_r(&time, &tm_storage);
 #endif
+	if (!tm) {
+		throw InternalException("FromTimeT failed: null pointer returned");
+	}
 
-	int32_t year = tm.tm_year + 1900;
-	int32_t month = tm.tm_mon + 1;
-	int32_t day = tm.tm_mday;
-	int32_t hour = tm.tm_hour;
-	int32_t min = tm.tm_min;
-	int32_t sec = tm.tm_sec;
+	int32_t year = tm->tm_year + 1900;
+	int32_t month = tm->tm_mon + 1;
+	int32_t day = tm->tm_mday;
+	int32_t hour = tm->tm_hour;
+	int32_t min = tm->tm_min;
+	int32_t sec = tm->tm_sec;
 
 	auto dt = Date::FromDate(year, month, day);
 	auto t = Time::FromTime(hour, min, sec, 0);