… .github/workflows/Main.yml
… src/include/duckdb/common/virtual_file_system.hpp
… src/include/duckdb/storage/external_file_cache.hpp
… src/optimizer/statistics/operator/propagate_aggregate.cpp
… src/storage/external_file_cache.cpp
… test/sql/copy/parquet/parallel_parquet_glob.test
… test/sql/tpch/tpch_hive_partition.test_slow