This PR modifies the Parquet reader to support `GetPartitionStats` when either (1) reading a single file, or (2) when all files that are read have their metadata cached. This allows us to emit partition data from Parquet files, also when globbing many files.

In order to avoid the cache invalidation from being expensive during this check, we perform the cache invalidation **only** using the `OpenFileInfo` objects that we have received from the multi file list - i.e. no additional file system calls are performed. This builds on the work in e.g. https://github.com/duckdb/duckdb-httpfs/pull/45 which alllows us to find the relevant information for performing cache invalidation during globbing and pushing that through all the relevant layers.

#### ListFilesExtended

In order to make this work also for local file systems, a new `ListFiles` method is introduced that returns `OpenFileInfo` objects instead of `string` objects:

```cpp
bool ListFiles(const string &directory, const std::function<void(OpenFileInfo &info)> &callback, 
               optional_ptr<FileOpener> opener = nullptr);
```

For backwards compatibility with other file systems, this method only needs to be optionally implemented through the `ListFilesExtended` method when `SupportsListFilesExtended` is overloaded to support true. This allows file systems to optionally implement either the extended callback or the older method.

```cpp
virtual bool ListFilesExtended(const string &directory,
                               const std::function<void(OpenFileInfo &info)> &callback,
                               optional_ptr<FileOpener> opener);
virtual bool SupportsListFilesExtended() const;
```

This PR then implements the `ListFilesExtended` callback for the `LocalFileSystem`:

* On Unix, we already did a `stat` call per file to find out if it was a directory - we now also read the file size (`st_size`) and last modified time (`st_mtime`) from the result.
* On Windows, globbing already returns a [`WIN32_FIND_DATAW`](https://learn.microsoft.com/en-us/windows/win32/api/minwinbase/ns-minwinbase-win32_find_dataw) struct for each file that contains the relevant information.


#### Performance

`GetPartitionStats` is currently only used for `COUNT(*)` (see https://github.com/duckdb/duckdb/pull/15301) - although there are plans to have this used in more scenarios.

##### Single File COUNT(*)

When reading a single file, we always emit `GetPartitionStats` also without metadata caching enabled - since we have already read the metadata as part of the initial schema detection step. This allows `COUNT(*)` to always be faster in this scenario using these changes:

```sql
SELECT COUNT(*) FROM hits.parquet;
```

|  main  |  new   |
|--------|--------|
| 0.054s | 0.035s |

##### Cached COUNT(*)

When enabling metadata caching, we can see that the cached metadata is used to accelarate `COUNT(*)`:

```sql
SET parquet_metadata_cache=true;
SELECT COUNT(*) FROM 'hits_partitioned/*.parquet';
```

| main  |  new   |
|-------|--------|
| 0.048 | 0.003s |