@Mytherin 2025-05-03
CC @Tishj @samansmink @Tmonster I think the `get_partition_stats` added in this PR might have to be set to `nullptr` for Iceberg/Delta (or rewritten/modified), since it does not take deletions into account. After this is merged I would at least add a test like to verify that this does not result in incorrect results:

```sql
SET parquet_metadata_cache=true;
SELECT COUNT(*) FROM table_with_deletions;
SELECT COUNT(*) FROM table_with_deletions;
```